BUG: Honor changes in urllib2 default handlers.
If the proxy handlers (or some other handlers) are changed in the
default urllibs opener (set via install_opener()), the
RemoteHTTPIndexedStore would not honor it because it cached the
opener in its `__init__` method. The `opener` has now been made a
property so that it always honors the changes to urllib2
default opener's handlers.